requirements.txt
attrs==17.4.0
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.4
django-markdownx==2.0.23
gunicorn==19.7.1
Markdown==2.6.11
more-itertools==4.1.0
Pillow==5.1.0
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pytest==3.5.0
pytest-django==3.2.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1

psycopg2==2.7.4